+++ src/components/indices/card.tsx
@@ -2,9 +2,9 @@ import * as React from "react"
 import cn from "classnames"
 import { Link } from "gatsby"
 import Image from "gatsby-image"
-import { AllBlogsQuery } from "../../types/graphql-types"
+import { AllBlogsQuery } from "../../../types/graphql-types"
 import styles from "./card.module.css"
-import { Tags } from "./tags"
+import { Tags } from "../tags"
 
 interface IProps {
   /** ブログコンテンツのfrontmatter */
