+++ src/templates/tagTemplate.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import { graphql } from "gatsby"
 import Layout from "../components/layout"
-import { Card } from "../components/card"
+import { Card } from "../components/indices/card"
 import styles from "./tagTemplate.module.css"
 import { TagTemplateQuery } from "../../types/graphql-types"
